[Ready for review] Update SearchAPI to search for foreach variable name and value. (#411)
* rename artifact search to search

* add backend api to search foreach

* update test

* add request identifier so that result of search requests are isolated

---------
