+++ docker-compose.development.yml
@@ -44,6 +44,7 @@ services:
       - FEATURE_HEARTBEAT_DISABLE=0
       - FEATURE_DB_LISTEN_DISABLE=0
       - FEATURE_ARTIFACT_SEARCH=1
+      - FEATURE_FOREACH_VAR_SEARCH=1
       - FEATURE_ARTIFACT_TABLE=1
       - CUSTOM_QUICKLINKS=$CUSTOM_QUICKLINKS
       - NOTIFICATIONS=$NOTIFICATIONS
