+++ services/ui_backend_service/api/__init__.py
@@ -2,7 +2,7 @@
 from .admin import AdminApi
 from .autocomplete import AutoCompleteApi
 from .artifact import ArtificatsApi
-from .artifactsearch import ArtifactSearchApi
+from .search import SearchApi
 from .dag import DagApi
 from .flow import FlowApi
 from .run import RunApi
